requirements.txt
cupy-cuda12x>=12.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
seaborn>=0.11.0
